--- a/src/Parallel/centrality/closeness.jl
+++ b/src/Parallel/centrality/closeness.jl
@@ -11,33 +11,10 @@ function closeness_centrality(
     end
 end
 
-function distr_closeness_centrality(
-    g::AbstractGraph, distmx::AbstractMatrix=weights(g); normalize=true
-)::Vector{Float64}
-    n_v = Int(nv(g))
-    closeness = SharedVector{Float64}(n_v)
-    fill!(closeness, 0.0)
-
-    @sync @distributed for u in vertices(g)
-        if degree(g, u) == 0     # no need to do Dijkstra here
-            closeness[u] = 0.0
-        else
-            d = Graphs.dijkstra_shortest_paths(g, u, distmx).dists
-            δ = filter(x -> x != typemax(x), d)
-            σ = sum(δ)
-            l = length(δ) - 1
-            if σ > 0
-                closeness[u] = l / σ
-                if normalize
-                    n = l * 1.0 / (n_v - 1)
-                    closeness[u] *= n
-                end
-            else
-                closeness[u] = 0.0
-            end
-        end
-    end
-    return sdata(closeness)
+function distr_closeness_centrality(args...; kwargs...)
+    return error(
+        "`parallel = :distributed` requested, but SharedArrays or Distributed is not loaded"
+    )
 end
 
 function threaded_closeness_centrality(