--- a/src/Parallel/distance.jl
+++ b/src/Parallel/distance.jl
@@ -4,7 +4,7 @@ function eccentricity(
     g::AbstractGraph,
     vs=vertices(g),
     distmx::AbstractMatrix{T}=weights(g);
-    parallel::Symbol=:distributed,
+    parallel::Symbol=:threads,
 ) where {T<:Number}
     return if parallel === :threads
         threaded_eccentricity(g, vs, distmx)
@@ -38,9 +38,7 @@ function threaded_eccentricity(
     return eccs
 end
 
-function eccentricity(
-    g::AbstractGraph, distmx::AbstractMatrix; parallel::Symbol=:distributed
-)
+function eccentricity(g::AbstractGraph, distmx::AbstractMatrix; parallel::Symbol=:threads)
     return eccentricity(g, vertices(g), distmx; parallel)
 end