--- a/src/Parallel/traversals/greedy_color.jl
+++ b/src/Parallel/traversals/greedy_color.jl
@@ -25,13 +25,10 @@ function threaded_random_greedy_color(g::AbstractGraph{T}, reps::Integer) where 
     return convert(Graphs.Coloring{T}, best)
 end
 
-function distr_random_greedy_color(g::AbstractGraph{T}, reps::Integer) where {T<:Integer}
-    best = @distributed (Graphs.best_color) for i in 1:reps
-        seq = shuffle(vertices(g))
-        Graphs.perm_greedy_color(g, seq)
-    end
-
-    return convert(Graphs.Coloring{T}, best)
+function distr_random_greedy_color(args...; kwargs...)
+    return error(
+        "`parallel = :distributed` requested, but SharedArrays or Distributed is not loaded"
+    )
 end
 
 function greedy_color(