--- a/src/Parallel/traversals/greedy_color.jl
+++ b/src/Parallel/traversals/greedy_color.jl
@@ -1,5 +1,5 @@
 function random_greedy_color(
-    g::AbstractGraph{T}, reps::Integer; parallel::Symbol=:distributed
+    g::AbstractGraph{T}, reps::Integer; parallel::Symbol=:threads
 ) where {T<:Integer}
     return if parallel === :threads
         threaded_random_greedy_color(g, reps)
@@ -32,10 +32,7 @@ function distr_random_greedy_color(args...; kwargs...)
 end
 
 function greedy_color(
-    g::AbstractGraph{U};
-    sort_degree::Bool=false,
-    reps::Integer=1,
-    parallel::Symbol=:distributed,
+    g::AbstractGraph{U}; sort_degree::Bool=false, reps::Integer=1, parallel::Symbol=:threads
 ) where {U<:Integer}
     return if sort_degree
         Graphs.degree_greedy_color(g)