This PR follows up on #429 to enable `parallel=:threads` by default and move Distributed-based parallelism to an extension, for faster loading times and reduced dependencies.

~~On Julia <1.9, the dependency is still there so this shouldn't break backwards compatibility (1.6 tests are passing for me locally)~~