@topolarity 2025-05-21
>   Expression: result ⊜ true
   Evaluated: ⟪result: 😭 FAILED: /Users/runner/work/Graphs.jl/Graphs.jl/Project.toml
    The file `/Users/runner/work/Graphs.jl/Graphs.jl/Project.toml` is not in canonical format.

Looks like 1.6 / 1.9 have a different notion of "canonical" for `[weakdeps]` entries in the Project.toml

I probably need to disable the test on Julia <1.9