@simonschoelly 2025-05-21
I don't want to be too negative here - but I am not sure if this is something we want right now. My reasons for this:
- It will make maintaining this package more difficult. It will also make testing it more difficult.
- It will make using it slightly more difficult as one has to explicitly import the relevant packages - unfortunately Julia's package manager does not have the abilities to declare features like Rust does for example.
- It is unclear how much this will increase loading time.
- Having no or very few dependencies was never the the goal of this package. In the past we removed some dependencies because some packages (like the linear optimization ones) were breaking this package during the transition from Julia v0.7 to v1.0.

If there is really the wish to have weak dependencies I am open to it - but first we should reach a consensus on that - so we should discus it in a Github issue.