@simonschoelly 2025-05-21
> > Expression: result ⊜ true
> > Evaluated: ⟪result: 😭 FAILED: /Users/runner/work/Graphs.jl/Graphs.jl/Project.toml
> > The file `/Users/runner/work/Graphs.jl/Graphs.jl/Project.toml` is not in canonical format.
> 
> Looks like 1.6 / 1.9 have a different notion of "canonical" for `[weakdeps]` entries in the Project.toml
> 
> I probably need to disable the test on Julia <1.9

Disabling tests for the lower bound version cannot be an option. We could raise the lower bounds to the latest long-term support release though - that would be v1.10. We can do that in a separate PR though.