@simonschoelly 2025-05-21
I think we are due for bumping the min version of this package anyway so I created a PR: #432 - let's see if someone objects.