@thchr 2025-05-22
> * It will make maintaining this package more difficult. It will also make testing it more difficult.

How will it become more difficult?

> * Having no or very few dependencies was never the the goal of this package. In the past we removed some dependencies because some packages (like the linear optimization ones) were breaking this package during the transition from Julia v0.7 to v1.0.

There's no downside to having fewer dependencies? The fewer dependencies Graphs.jl has, the lower the barrier is to some other package deciding to take a dependency on Graphs.jl.