@simonschoelly 2025-05-23
> > * It is unclear how much this will increase loading time.
> 
> Wouldn't this strictly decrease loading time? And decrease compilation time as well? It is a small change for this specific package, but for packages that depend on Graphs.jl (and other packages) this can start adding up. It is why DiffEq.jl split out all of its solvers in separate packages.
> 
> e.g. running `julia> @time @eval using Graphs` on 1.11.5
> 
> master: `0.540166 seconds (411.04 k allocations: 26.379 MiB, 5.20% gc time, 15.12% compilation time)`
> 
> this pr: `0.502729 seconds (380.79 k allocations: 24.282 MiB, 4.67% gc time, 8.19% compilation time)`
> 
> > * Having no or very few dependencies was never the the goal of this package.
> 
> But happening to have few dependencies makes the life of downstream package maintainers betters. If Graphs unconditionally depends on Distributed, then other packages that have extensions for Distributed will run precompilation for the extensions. Changes like this are overall beneficial for the user because way less unnecessary precompilation will be happening.

Are you running this on some kind of beefed up cluster? For me it takes around 6 to 7 seconds to import `Graphs.jl` after switching my branch.

Also, what does `@eval` do here?

I have not figured out how to invalidate all precompiled code - but when running this multiple times the loading time varies by quite a bit and sometimes on branch is faster and the other time the other. So unless you repeated this experiment multiple times I am not sure if we really have a speed up of 7.4%.