@Krastanov 2025-05-23
The runs above were on a cold laptop that was top-of-the-line in 2019, in performance mode, and are admittedly difficult to reproduce. The runs below are on a well-cooled steady-state desktop that is not throttling. The difference between what the two of us are getting above are probably loading vs precompiling, so below I separate those.

On julia 1.12-beta3

Either way though, this testing should not matter much -- putting things in extensions inherently leads to less precompilation and less loading. The improvements can be very minimal (and they are negligible in this particular case), but the improvements should be unconditional. More importantly, this is particularly useful for downstream libraries that depend of Graphs. I like the slow push for more extensions and conditional dependencies in the ecosystem as a whole, because that makes my life as a developer of a library with many dependencies easier.

## loading time comparison (no precompilation): very minor improvement in timing and allocations

this PR

```
> for i in (seq 1 10); julia --project=. -t1,1 -e "@time @eval using Graphs"; end
  0.358056 seconds (504.41 k allocations: 32.222 MiB, 7.62% compilation time)
  0.359722 seconds (504.41 k allocations: 32.222 MiB, 7.71% compilation time)
  0.354502 seconds (504.41 k allocations: 32.222 MiB, 7.75% compilation time)
  0.360770 seconds (504.41 k allocations: 32.237 MiB, 7.67% compilation time)
  0.361303 seconds (504.41 k allocations: 32.237 MiB, 7.52% compilation time)
  0.354667 seconds (504.41 k allocations: 32.221 MiB, 7.97% compilation time)
  0.361170 seconds (504.41 k allocations: 32.221 MiB, 7.56% compilation time)
  0.359616 seconds (504.41 k allocations: 32.221 MiB, 7.68% compilation time)
  0.372852 seconds (504.41 k allocations: 32.222 MiB, 7.37% compilation time)
  0.359655 seconds (504.41 k allocations: 32.221 MiB, 7.62% compilation time)
```

master:

```
> for i in (seq 1 10); julia --project=. -t1,1 -e "@time @eval using Graphs"; end
  0.379161 seconds (531.04 k allocations: 34.046 MiB, 7.34% compilation time)
  0.371243 seconds (531.04 k allocations: 34.045 MiB, 7.51% compilation time)
  0.373846 seconds (531.04 k allocations: 34.046 MiB, 7.47% compilation time)
  0.372231 seconds (531.04 k allocations: 34.029 MiB, 7.34% compilation time)
  0.372268 seconds (531.04 k allocations: 34.031 MiB, 7.38% compilation time)
  0.362420 seconds (531.04 k allocations: 34.031 MiB, 7.64% compilation time)
  0.362064 seconds (531.04 k allocations: 34.030 MiB, 7.99% compilation time)
  0.369258 seconds (531.04 k allocations: 34.046 MiB, 7.38% compilation time)
  0.373228 seconds (531.04 k allocations: 34.030 MiB, 7.44% compilation time)
  0.372276 seconds (531.04 k allocations: 34.030 MiB, 7.44% compilation time)
```

## precompilation of the entire environment: very minor improvement in allocations, negligible improvement in timing

this pr:

```
> for i in (seq 1 10); rm -rf ~/.julia/compiled/v1.12; julia --project=. -t1,1 -e "@time @eval using Graphs" 2> /dev/null; end
 10.375091 seconds (645.68 k allocations: 47.397 MiB, 0.81% gc time, 0.41% compilation time)
 10.369979 seconds (645.81 k allocations: 47.400 MiB, 0.80% gc time, 0.41% compilation time)
 10.407871 seconds (645.76 k allocations: 47.446 MiB, 0.80% gc time, 0.40% compilation time)
 10.409445 seconds (645.81 k allocations: 47.377 MiB, 0.80% gc time, 0.41% compilation time)
 10.371197 seconds (645.68 k allocations: 47.375 MiB, 0.81% gc time, 0.41% compilation time)
 10.450528 seconds (645.60 k allocations: 47.384 MiB, 0.79% gc time, 0.40% compilation time)
 10.317081 seconds (645.69 k allocations: 47.455 MiB, 0.81% gc time, 0.41% compilation time)
 10.446430 seconds (645.61 k allocations: 47.395 MiB, 0.80% gc time, 0.41% compilation time)
 10.371907 seconds (645.68 k allocations: 47.403 MiB, 0.80% gc time, 0.44% compilation time)
 10.448137 seconds (645.85 k allocations: 47.467 MiB, 0.79% gc time, 0.41% compilation time)
```

master:

```
> for i in (seq 1 10); rm -rf ~/.julia/compiled/v1.12; julia --project=. -t1,1 -e "@time @eval using Graphs" 2> /dev/null; end
 10.422142 seconds (679.91 k allocations: 50.342 MiB, 0.79% gc time, 0.42% compilation time)
 10.335160 seconds (680.00 k allocations: 50.288 MiB, 0.81% gc time, 0.41% compilation time)
 10.314185 seconds (680.08 k allocations: 50.408 MiB, 0.80% gc time, 0.41% compilation time)
 10.343851 seconds (679.84 k allocations: 50.287 MiB, 0.83% gc time, 0.41% compilation time)
 10.635309 seconds (679.92 k allocations: 50.347 MiB, 0.79% gc time, 0.40% compilation time)
 10.638776 seconds (680.08 k allocations: 50.298 MiB, 0.80% gc time, 0.41% compilation time)
 10.688697 seconds (679.84 k allocations: 50.419 MiB, 0.78% gc time, 0.43% compilation time)
 10.375991 seconds (680.08 k allocations: 50.428 MiB, 0.80% gc time, 0.41% compilation time)
 10.341547 seconds (679.84 k allocations: 50.296 MiB, 0.80% gc time, 0.41% compilation time)
 10.335641 seconds (680.04 k allocations: 50.353 MiB, 0.82% gc time, 0.46% compilation time)
```

## precompilation of just Graphs.jl: no improvement

this pr:

```
> for i in (seq 1 10); echo "#" >> Graphs.jl/src/Graphs.jl; julia --project=. -t1,1 -e "@time @eval using Graphs" 2> /dev/null; end
  3.595370 seconds (633.06 k allocations: 45.822 MiB, 2.32% gc time, 1.18% compilation time)
  3.546514 seconds (633.06 k allocations: 45.759 MiB, 2.34% gc time, 1.18% compilation time)
  3.584489 seconds (633.06 k allocations: 45.822 MiB, 2.29% gc time, 1.17% compilation time)
  3.566042 seconds (633.06 k allocations: 45.760 MiB, 2.33% gc time, 1.18% compilation time)
  3.560505 seconds (633.06 k allocations: 45.759 MiB, 2.34% gc time, 1.16% compilation time)
  3.573339 seconds (633.06 k allocations: 45.760 MiB, 2.34% gc time, 1.17% compilation time)
  3.583969 seconds (633.06 k allocations: 45.776 MiB, 2.36% gc time, 1.19% compilation time)
  3.568351 seconds (633.06 k allocations: 45.761 MiB, 2.33% gc time, 1.17% compilation time)
  3.598015 seconds (633.06 k allocations: 45.756 MiB, 2.62% gc time, 1.19% compilation time)
  3.525711 seconds (633.06 k allocations: 45.761 MiB, 2.37% gc time, 1.19% compilation time)
```

master:

```
> for i in (seq 1 10); echo "#" >> Graphs.jl/src/Graphs.jl; julia --project=. -t1,1 -e "@time @eval using Graphs" 2> /dev/null; end
  3.555798 seconds (633.07 k allocations: 45.842 MiB, 2.33% gc time, 1.19% compilation time)
  3.575576 seconds (633.06 k allocations: 45.818 MiB, 2.33% gc time, 1.19% compilation time)
  3.589061 seconds (633.07 k allocations: 45.839 MiB, 2.32% gc time, 1.18% compilation time)
  3.546660 seconds (633.06 k allocations: 45.765 MiB, 2.32% gc time, 1.19% compilation time)
  3.579585 seconds (633.06 k allocations: 45.759 MiB, 2.31% gc time, 1.21% compilation time)
  3.562644 seconds (633.06 k allocations: 45.843 MiB, 2.34% gc time, 1.18% compilation time)
  3.563017 seconds (633.06 k allocations: 45.780 MiB, 2.35% gc time, 1.20% compilation time)
  3.604094 seconds (633.06 k allocations: 45.822 MiB, 2.42% gc time, 1.18% compilation time)
  3.566834 seconds (633.06 k allocations: 45.777 MiB, 2.35% gc time, 1.20% compilation time)
  3.540509 seconds (633.06 k allocations: 45.826 MiB, 2.38% gc time, 1.24% compilation time)
```

The `@eval` is necessary to make sure all compilation work is included in the timing. I do not know the details of why, but this PR can be instructive if you want to delve deeper https://github.com/JuliaLang/julia/pull/58015